Remove unused files and clean up project structure
