--- progression_analysis.md
@@ -1,126 +1,0 @@
-# Bot Progression Analysis - 13-15 Day Timeline
-
-## Economic Changes Made
-
-### Starting Money (Reduced by ~85%)
-- USA: 5B → 800K
-- Russia: 3B → 650K  
-- China: 4B → 720K
-- Iran: 2B → 500K
-- Germany: 3.5B → 680K
-- Japan: 4B → 720K
-- UK: 3.5B → 680K
-- France: 3B → 620K
-- India: 1.5B → 450K
-- South Korea: 3B → 620K
-- Israel: 2B → 520K
-
-### Income Per Hour (Reduced by ~93%)
-**Example - Iran:**
-- Oil exports: 40M → 280K (×1.3 = 364K)
-- Regional trade: 20M → 160K (×1.5 = 240K) 
-- Cultural exports: 10M → 70K (×1.2 = 84K)
-- **Total per hour: ~688K** (with multipliers)
-
-### Message Rewards (Reduced by ~93%)
-- Per message: 50K → 3.5K
-- Level bonus: 5M → 400K per level
-
-### Technology Costs (Reduced by ~97%)
-- Military Engineering: 50M → 1.2M
-- Aerodynamics: 80M → 1.8M
-- Nuclear Technology: 200M → 4.5M
-- Quantum Computing: 300M → 6.8M
-- Nuclear Weapons: 500M → 12M
-
-### Military Assets (Reduced by ~70-85%)
-**Basic Units:**
-- Militia: 50K → 15K
-- Infantry: 100K → 25K
-- Marines: 250K → 60K
-
-**Advanced Units:**
-- F-22 Raptor: 300M → 85M
-- F-35 Lightning: 500M → 120M
-- B-2 Spirit: 800M → 180M
-
-**Nuclear Weapons:**
-- Minuteman: 200M → 55M
-- Trident: 300M → 82M
-
-## Progression Timeline Calculation
-
-### Daily Income Potential
-**Iran Example (Best oil income):**
-- Base hourly: ~688K
-- Daily (24 hours): ~16.5M
-- With level/tech bonuses: ~20M/day by mid-game
-
-### Major Milestones
-
-**Days 1-3: Basic Setup**
-- Select country, start with 500K-800K
-- Buy basic infantry, research first techs
-- Daily income: ~10-15M
-
-**Days 4-7: Mid-tier Development**  
-- Research advanced metallurgy (2.8M)
-- Buy country-specific tanks (4-9M each)
-- Research aerodynamics (1.8M)
-- Daily income: ~15-20M
-
-**Days 8-12: Advanced Military**
-- Research stealth technology (3.5M)
-- Buy stealth aircraft (85-125M each)
-- Research nuclear technology (4.5M)
-- Daily income: ~20-25M
-
-**Days 13-15: End Game**
-- Research quantum computing (6.8M)
-- Research nuclear weapons (12M)
-- Buy nuclear missiles (55-95M each)
-- Buy future tech (40-85M each)
-
-### Country-Specific Examples
-
-**Iran Players:**
-- Unique units: Sejjil (800K), Emad (950K), Khorramshahr (1.2M)
-- Shahed-136 drones (25K) - Very affordable
-- Karrar tank (4.8M), Zulfiqar tank (3.2M)
-- Quds Force (320K), Takavar (280K)
-
-**USA Players:**
-- Exclusive access to most expensive units
-- F-22 Raptor (85M), F-35 (120M), B-2 (180M)
-- Nuclear arsenal: Minuteman (55M), Trident (82M)
-- Elite forces: Navy SEALs (300K), Delta Force (450K)
-
-**Russia Players:**  
-- T-14 Armata (6.8M), Su-57 (95M)
-- Hypersonic Glider (75M), Pak DA (380M)
-- Nuclear: Topol-M (65M), Sarmat (95M)
-- Spetsnaz (380K)
-
-## Verification
-
-**Total Cost for Full Arsenal:** ~800M - 1.2B per country
-**Total Income over 15 days:** ~300-400M
-**Conclusion:** Players will need 13-15 days of consistent play to achieve top-tier military capabilities, making the progression appropriately challenging while maintaining engagement.
-
-## New Features Added
-
-### Country-Specific Units
-- Each country now has exclusive military units
-- 120+ new military assets across all categories
-- Realistic country restrictions (e.g., only Iran can buy Sejjil missiles)
-
-### Enhanced Categories
-- **Infantry:** 12 country-specific special forces
-- **Armored:** 12 country-specific main battle tanks  
-- **Aircraft:** 15+ country-specific fighters and stealth aircraft
-- **Naval:** 12+ country-specific ships and submarines
-- **Missiles:** 15+ country-specific missile systems
-- **Artillery:** 12+ country-specific artillery systems
-- **Advanced Tech:** 10+ country-specific future weapons
-
-This creates a much more diverse and strategic gameplay experience where country choice significantly impacts available military options.
