--- requirements.txt
@@ -1,7 +1,0 @@
-python-bale-bot>=1.0.0
-asyncio
-json
-datetime
-random
-traceback
-time
